LGVISIUM-102: create Sidebar Extractor classes
