+++ src/stratigraphy/sidebar/__init__.py
@@ -1,9 +1,19 @@
 """Modules for Sidebars, representing depths or other data displayed to the side of material descriptions."""
 
 from .a_above_b_sidebar import AAboveBSidebar
+from .a_above_b_sidebar_extractor import AAboveBSidebarExtractor
 from .a_above_b_sidebar_validator import AAboveBSidebarValidator
 from .a_to_b_sidebar import AToBSidebar
+from .a_to_b_sidebar_extractor import AToBSidebarExtractor
 from .layer_identifier_sidebar import LayerIdentifierSidebar
 from .sidebar import Sidebar
 
-__all__ = ["Sidebar", "AAboveBSidebar", "AToBSidebar", "LayerIdentifierSidebar", "AAboveBSidebarValidator"]
+__all__ = [
+    "Sidebar",
+    "AAboveBSidebar",
+    "AAboveBSidebarExtractor",
+    "AAboveBSidebarValidator",
+    "AToBSidebar",
+    "AToBSidebarExtractor",
+    "LayerIdentifierSidebar",
+]
